requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncpg
python-dotenv
Jinja2